Move db credential interface and implementations to shared sub-project
