+++ shared/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/database/ConnectionPool.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.wrappers;
+package gov.hhs.cdc.trustedintermediary.wrappers.database;
 
 import java.sql.Connection;
 import java.sql.SQLException;
