+++ shared/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/database/DatabaseCredentialsProvider.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.external.database;
+package gov.hhs.cdc.trustedintermediary.wrappers.database;
 
 /** This interface represents a provider for retrieving database credentials. */
 public interface DatabaseCredentialsProvider {
